+++ src/main/java/sci2s/sparkfingerprint/SparkMatcherJiang.scala
@@ -161,7 +161,7 @@ object SparkMatcherJiang {
 
   def computeScores5(
       templateLS : RDD[(String, LocalStructureJiang)],
-      inputLS : Broadcast[Array[(String, Array[LocalStructureJiang])]]) : RDD[(String, (String, Double))] = {
+      inputLS : Broadcast[Array[(String, Array[LocalStructureJiang])]]) : RDD[(String, (String, Float))] = {
       
     // First, compute the partial scores of each template LS with each input fingerprint.
     val scores = templateLS.groupByKey().flatMapValues({ tlsarray =>
